Rename dir - 3 associated_objects_management
