+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/associated_object_management/accelerations_recommendation_callout.test.tsx
