+++ src/plugins/data_source_management/public/plugin.ts
@@ -49,7 +49,7 @@ import {
 } from '../framework/types';
 import { AccelerationDetailsFlyout } from './components/direct_query_data_sources_components/acceleration_management/acceleration_details_flyout';
 import { CreateAcceleration } from './components/direct_query_data_sources_components/acceleration_creation/create/create_acceleration';
-import { AssociatedObjectsDetailsFlyout } from './components/direct_query_data_sources_components/direct_query_associated_object_management/associated_objects_details_flyout';
+import { AssociatedObjectsDetailsFlyout } from './components/direct_query_data_sources_components/associated_object_management/associated_objects_details_flyout';
 
 export const [
   getRenderAccelerationDetailsFlyout,
